pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "andrewcli"
version = "0.1.0"
description = "A lightweight, fully async Python agent with rolling memory and LLM-based routing"
license = { text = "GPL-3.0" }
requires-python = ">=3.10"
dependencies = [
    "openai",
    "PyYAML",
    "fastapi",
    "uvicorn[standard]",
    "pydantic",
    "PyQt6",
    "selenium>=4.15.0",
    "mcp[cli]",
    "httpx",
]

[project.scripts]
andrewcli = "andrewcli:main"

[tool.setuptools]
py-modules = ["andrewcli"]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.setuptools.package-data]
# Bundled defaults are seeded into ~/.config/andrewcli/ on first run
# (see src/shared/paths.py::_seed_user_config). Ship every file under
# src/defaults/ so domains, events, the global config.yaml, skills,
# tools, and __init__.py markers all survive a regular pip install.
"src" = ["defaults/**/*", "defaults/**/.*"]
"src.tray"  = ["*.css"]
